Finish RouteOptionsTableViewCell
