+++ TCAT/Utilities/Analytics.swift
@@ -18,8 +18,6 @@ class Analytics {
         #if !DEBUG
             let fabricEvent = payload.convertToFabric()
             Answers.logCustomEvent(withName: fabricEvent.name, customAttributes: fabricEvent.attributes)
-
-            print("Logging \(fabricEvent.name):", fabricEvent.attributes ?? [:])
         #endif
     }
 }
